+++ src/models/PanelSettings.ts
@@ -5,14 +5,22 @@ export interface PanelSettings {
   tags: string[];
   categories: string[];
   freeform: boolean;
+  scripts: CustomScript[];
 }
 
 export interface SEO {
   title: number;
   description: number;
+  descriptionField: string;
 }
 
 export interface Slug {
   prefix: number;
   suffix: number;
+}
+
+export interface CustomScript {
+  title: string;
+  script: string;
+  nodeBin?: string;
 }