+++ src/viewpanel/CommandToCode.ts
@@ -2,9 +2,14 @@ export enum CommandToCode {
   getData = "get-data",
   updateSlug = 'update-slug',
   updateDate = 'update-date',
+  updateLastMod = 'update-lastmod',
   publish = 'publish',
   updateTags = "update-tags",
   updateCategories = "update-categories",
   addTagToSettings = "add-tag",
-  addCategoryToSettings = "add-category"
+  addCategoryToSettings = "add-category",
+  openSettings = "open-settings",
+  openFile = "open-file",
+  openProject = "open-project",
+  runCustomScript = "custom-script"
 }