+++ src/viewpanel/components/Actions.tsx
@@ -1,6 +1,8 @@
 import * as React from 'react';
 import { PanelSettings } from '../../models/PanelSettings';
+import { CustomScript } from './CustomScript';
 import { DateAction } from './DateAction';
+import { Icon } from './Icon';
 import { PublishAction } from './PublishAction';
 import { SlugAction } from './SlugAction';
 
@@ -17,14 +19,22 @@ export const Actions: React.FunctionComponent<IActionsProps> = (props: React.Pro
   }
 
   return (
-    <div className={`article__actions`}>
-      <h3>Actions</h3>
+    <div className={`section article__actions`}>
+      <h3><Icon name={`rocket`} /> Actions</h3>
 
       { metadata && metadata.title && <SlugAction value={metadata.title} crntValue={metadata.slug} slugOpts={settings.slug} /> }
       
       <DateAction />
 
       { metadata && typeof metadata.draft !== undefined && <PublishAction draft={metadata.draft} />}
+
+      {
+        (settings && settings.scripts && settings.scripts.length > 0) && (
+          settings.scripts.map((value) => (
+            <CustomScript key={value.title.replace(/ /g, '')} {...value} />
+          ))
+        )
+      }
     </div>
   );
 };