+++ src/viewpanel/components/Icon.tsx
@@ -1,0 +1,10 @@
+import * as React from 'react';
+
+export interface IIconProps {
+  name: string;
+}
+
+export const Icon: React.FunctionComponent<IIconProps> = ({ name }: React.PropsWithChildren<IIconProps>) => {
+  
+  return (<i className={`codicon codicon-${name}`}></i>);
+};