+++ src/viewpanel/components/SeoStatus.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
 import { SEO } from '../../models/PanelSettings';
+import { Icon } from './Icon';
 import { SeoDetails } from './SeoDetails';
 
 export interface ISeoStatusProps {
@@ -9,17 +10,21 @@ export interface ISeoStatusProps {
 
 export const SeoStatus: React.FunctionComponent<ISeoStatusProps> = (props: React.PropsWithChildren<ISeoStatusProps>) => {
   const { data, seo } = props;
-  const { title, description } = data;
+  const { title } = data;
 
-  if (!title && !description) {
+  const { descriptionField } = seo;
+
+  if (!title && !data[descriptionField]) {
     return null;
   }
 
   return (
-    <div className="seo__status">
-      <h3>SEO Status</h3>
+    <div className="section seo__status">
+      <h3>
+        <Icon name="search" /> SEO Status
+      </h3>
       { (title && seo.title > 0) && <SeoDetails title="Title" allowedLength={seo.title} value={title} /> }
-      { (description && seo.description > 0) && <SeoDetails title="Description" allowedLength={seo.description} value={description} /> }
+      { (data[descriptionField] && seo.description > 0) && <SeoDetails title="Description" allowedLength={seo.description} value={data[descriptionField]} /> }
     </div>
   );
 };